Fix handling of context in macros, etc.
There was a bug in the handling of context objects when passing between
methods of the Machine and Expression classes.  Before passing context to
Expression, the local context would be merged with the global context.
If `Expression.execute()` had to defer to Machine again (as is the case
with, e.g.,  directives nested inside of macro declarations), the
Machine class would now have a copy of all global variables in its local
context.  Then, for example, a @set would operate on the global context,
as expected, but a "get" with `@{varName}` would get the unaltered copy
from the local/merged context object.